Still trying to fix paging, i'm hoping this is whats causing the VFS problem so i can fix two issues at once

+++ include/memory.h
@@ -45,7 +45,7 @@ struct	boundary_tag
 	struct boundary_tag *prev;	//< Linked list info.
 };
 
-void init_paging(uint32_t total_frames, uint32_t ident_addr, uint32_t ident_len);
+void init_paging(uint64_t total_frames, uint64_t ident_addr, uint64_t ident_len);
 extern void paging_enable();
 struct page *get_page(uint32_t address, int make, struct page_directory *dir);
 struct free_header *find_block(uint32_t size, uint8_t align);
